chore: set up the environment
