+++ .eslintrc
@@ -1,0 +1,37 @@
+{
+    "parser": "@typescript-eslint/parser",
+    "parserOptions": {
+      "project": "tsconfig.json",
+      "sourceType": "module"
+    },
+    "plugins": ["@typescript-eslint/eslint-plugin", "prettier"],
+    "extends": [
+      "plugin:@typescript-eslint/recommended",
+      "plugin:prettier/recommended",
+      "prettier"
+    ],
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "ignorePatterns": [],
+    "rules": {
+      "prettier/prettier": "error",
+      "no-var": "error",
+      "semi": "error",
+      "indent": ["error", 2, { "SwitchCase": 1 }],
+      "no-multi-spaces": "error",
+      "space-in-parens": "error",
+      "no-multiple-empty-lines": "error",
+      "prefer-const": "error",
+      "no-undef": "off",
+      "no-case-declarations": "off",
+      "no-unused-vars": [
+        "error",
+        {
+          "varsIgnorePattern": "^[A-Z]"
+        }
+      ]
+    }
+  }
+
